plugins/filestorage/src/main/java/pt/ua/ieeta/filestorageplugin/utils/PathTranslator.java
/**
 * Copyright (C) 2015  Universidade de Aveiro, DETI/IEETA, Bioinformatics Group - http://bioinformatics.ua.pt/
 *
 * This file is part of Dicoogle/filestorage.
 *
 * Dicoogle/filestorage is free software: you can redistribute it and/or modify
 * it under the terms of the GNU General Public License as published by
 * the Free Software Foundation, either version 3 of the License, or
 * (at your option) any later version.
 *
 * Dicoogle/filestorage is distributed in the hope that it will be useful,
 * but WITHOUT ANY WARRANTY; without even the implied warranty of
 * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 * GNU General Public License for more details.
 *
 * You should have received a copy of the GNU General Public License
 * along with Dicoogle.  If not, see <http://www.gnu.org/licenses/>.
 */
package pt.ua.ieeta.filestorageplugin.utils;

import java.net.URI;
import java.nio.file.Path;


public interface PathTranslator {

    /** Given the item URI, translate it to its path in the file system.
     * @param uri the effective URI of the stored item.
     */
    public Path toPath(URI uri);

    /** Given the file system path to the URI, translate it to its URI.
     * @param path the path to the file.
     */
    public URI toUri(Path path);
}
